requirements.txt
PyQt6>=6.6
firebase-admin>=6.5.0
python-dotenv>=1.0.1
google-cloud-firestore>=2.16.0
qrcode[pil]>=7.4
Pillow>=10.3.0
python-barcode>=0.15.1
python-escpos>=3.1
pyusb>=1.2.1
pyserial>=3.5
